requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
oauthlib==3.2.2
Pillow==9.3.0
py-cord==2.3.2
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
